feat(GoogleDocs Ad): move to promotions.jsx and use React.memo
