+++ static/js/Misc.jsx
@@ -23,8 +23,6 @@ import Cookies from "js-cookie";
 import {EditTextInfo} from "./BookPage";
 import ReactMarkdown from 'react-markdown';
 import TrackG4 from "./sefaria/trackG4";
-import {trackSidebarAdClick, trackSidebarAdImpression} from "./Promotions";
-const cookie = Sefaria._inBrowser ? $.cookie : Sefaria.util.cookie;
 
 /**
  * Component meant to simply denote a language specific string to go inside an InterfaceText element
@@ -62,35 +60,7 @@ const __filterChildrenByLanguage = (children, language) => {
   let newChildren = chlArr.filter(x=> x.type == currLangComponent);
   return newChildren;
 };
-const GDocAdvertText = () => {
-    return    <InterfaceText>
-                <EnglishText> Add texts directly to your Google Docs with our <span id="newExtension">new extension</span>! <a href="https://sefaria.org/sheets/529099?origin=AddToSheetsPromo">Learn more</a></EnglishText>
-                <HebrewText> הוסיפו טקסטים מספריא ישירות לקובץ גוגל עם <span id="newExtension">התוסף החדש</span> שלנו! <a href="https://sefaria.org/sheets/529099?origin=AddToSheetsPromo">למדו עוד</a></HebrewText>
-             </InterfaceText>;
-}
-const GDocAdvertBox = () => {
-    const firstRender = useRef(true);
-    useEffect(() => {
-        if (firstRender.current) {
-          firstRender.current = false;
-          if (!cookie(gdocInstalled)) { // on first render and "Install Now" has never been clicked before
-              trackSidebarAdImpression(gdocsCampaign);
-          }
-        }
-      })
-    const gdocsCampaign = {campaignId: "GDocs_Promo_AddToSheet"};
-    const gdocInstalled = 'gdoc_installed';
-    const handleInstall = () => {
-        cookie(gdocInstalled, JSON.stringify(1), {path: "/"});
-        trackSidebarAdClick(gdocsCampaign);
-    }
-    return !cookie(gdocInstalled) &&
-            <div className="gDocAdvertBox">
-              <GDocAdvertText/>
-              <div id="installNow"><a href={'https://workspace.google.com/marketplace/app/sefaria/849562338091?utm_source=SefariaOrg&utm_medium=SidebarAdButton&utm_campaign=AddToSheetPromotion&utm_content=InstallFromAddToSheet'}
-                                      onClick={handleInstall}><InterfaceText>Install Now</InterfaceText></a></div>
-            </div>;
-}
+
 
 const InterfaceText = ({text, html, markdown, children, context, disallowedMarkdownElements=[]}) => {
   /**
@@ -3375,6 +3345,5 @@ export {
   requestWithCallBack,
   OnInView,
   TopicPictureUploader,
-  ImageWithCaption,
-  GDocAdvertBox
+  ImageWithCaption
 };
